refactored server.ts into app.ts and server.ts
for testing and clarity the server and app were separated into two
different files
