+++ app/auth/jwt-auth.middleware.ts
@@ -1,6 +1,6 @@
 import {Request, Response, NextFunction} from "express";
 import {APIClientInfo, JWTAuthService} from "./jwt-auth.service";
-import {app} from "../server";
+import {app} from "../app";
 import {API_JWT_SECRET} from "./auth.constants";
 
 // This middleware will extend the request object
