+++ app/controllers/api.controller.ts
@@ -1,5 +1,5 @@
 import {NextFunction, Request, Response, Router} from "express";
-import {Error} from "../server";
+import {Error} from "../app";
 import {ChargeEvent, ChargeEventBase, CheckIn, CKCheckIn, Ladelog} from "../models/chargeevent.model";
 import {AppRequest, jwtAuth} from "../auth/jwt-auth.middleware";
 import {Model, ValidationError} from "mongoose";
